Prawie skończone

+++ frontend/app/src/main/java/com/example/magazyn/api/dtos/StanMagazynuDTO.kt
@@ -9,6 +9,6 @@ import com.google.gson.annotations.SerializedName
 data class StanMagazynuDTO(
     val id: Int,
     @SerializedName("ilosc")
-    val ilosc: Double,
+    val ilosc: Int,
     val jednostka: String
 )
